+++ firebase-vscode/webviews/components/EmulatorPanel.tsx
@@ -7,18 +7,18 @@ import {
   VSCodeTextField,
 } from "@vscode/webview-ui-toolkit/react";
 import React, { useState } from "react";
-import { Spacer } from "./components/ui/Spacer";
-import { broker } from "./globals/html-broker";
-import { PanelSection } from "./components/ui/PanelSection";
-import { FirebaseConfig } from "../../src/firebaseConfig";
+import { Spacer } from "./ui/Spacer";
+import { broker } from "../globals/html-broker";
+import { PanelSection } from "./ui/PanelSection";
+import { FirebaseConfig } from "../../../src/firebaseConfig";
 import {
   RunningEmulatorInfo,
   EmulatorUiSelections,
-} from "../common/messaging/types";
+} from "../../common/messaging/types";
 import { VSCodeDropdown } from "@vscode/webview-ui-toolkit/react";
 import { VSCodeOption } from "@vscode/webview-ui-toolkit/react";
-import { EmulatorInfo } from "../../src/emulator/types";
-import { webLogger } from "./globals/web-logger";
+import { EmulatorInfo } from "../../../src/emulator/types";
+import { webLogger } from "../globals/web-logger";
 
 const DEFAULT_EMULATOR_UI_SELECTIONS: EmulatorUiSelections = {
   projectId: "demo-something",
@@ -33,14 +33,20 @@ const DEFAULT_EMULATOR_UI_SELECTIONS: EmulatorUiSelections = {
  */
 export function EmulatorPanel({
   firebaseJson,
+  projectId,
 }: {
   firebaseJson: FirebaseConfig;
+  projectId?: string | undefined;
 }) {
   if (!firebaseJson) {
     throw Error("Expected a valid FirebaseConfig.");
   }
+  const defaultState = DEFAULT_EMULATOR_UI_SELECTIONS;
+  if (projectId) {
+    defaultState.projectId = getProjectIdForMode(projectId, defaultState.mode);
+  }
   const [emulatorUiSelections, setEmulatorUiSelections] =
-    useState<EmulatorUiSelections>(DEFAULT_EMULATOR_UI_SELECTIONS);
+    useState<EmulatorUiSelections>(defaultState);
 
   webLogger.debug(
     "initial state ui selections:" + JSON.stringify(emulatorUiSelections)
@@ -74,9 +80,14 @@ export function EmulatorPanel({
   });
 
   broker.on("notifyEmulatorImportFolder", ({ folder }) => {
-    webLogger.debug(`notifyEmulatorImportFolder received in sidebar: ${folder}`);
-    emulatorUiSelections.importStateFolderPath = folder;
-    setEmulatorUiSelectionsAndSaveToWorkspace({ ...emulatorUiSelections }); // rerender clone
+    webLogger.debug(
+      `notifyEmulatorImportFolder received in sidebar: ${folder}`
+    );
+    const newSelections = {
+      ...emulatorUiSelections,
+      importStateFolderPath: folder,
+    };
+    setEmulatorUiSelectionsAndSaveToWorkspace(newSelections);
   });
 
   function launchEmulators() {
@@ -137,15 +148,22 @@ export function EmulatorPanel({
 
   function emulatorModeChanged(event: React.ChangeEvent<HTMLSelectElement>) {
     webLogger.debug("emulatorModeChanged: " + event.target.value);
-    const selections: EmulatorUiSelections = emulatorUiSelections;
-    selections.mode = event.target.value as typeof emulatorUiSelections.mode;
-    setEmulatorUiSelectionsAndSaveToWorkspace(selections);
+    const newSelections: EmulatorUiSelections = { ...emulatorUiSelections };
+    newSelections.mode = event.target.value as typeof emulatorUiSelections.mode;
+    newSelections.projectId = getProjectIdForMode(
+      projectId,
+      newSelections.mode
+    );
+    setEmulatorUiSelectionsAndSaveToWorkspace(newSelections);
   }
 
   function clearImportFolder() {
     console.log(`clearImportFolder`);
-    emulatorUiSelections.importStateFolderPath = "";
-    setEmulatorUiSelectionsAndSaveToWorkspace({ ...emulatorUiSelections });
+    const newSelections = {
+      ...emulatorUiSelections,
+      importStateFolderPath: "",
+    };
+    setEmulatorUiSelectionsAndSaveToWorkspace(newSelections);
   }
 
   // Make it pretty for the screen. Filter out the logging emulator since it's
@@ -208,7 +226,9 @@ export function EmulatorPanel({
         onChange={(event) => emulatorModeChanged(event)}
       >
         <VSCodeOption value="all">All emulators</VSCodeOption>
-        {!!firebaseJson.hosting && <VSCodeOption value="hosting">Only hosting</VSCodeOption>}
+        {!!firebaseJson.hosting && (
+          <VSCodeOption value="hosting">Only hosting</VSCodeOption>
+        )}
       </VSCodeDropdown>
       {runningEmulatorInfo ? (
         <>
@@ -246,4 +266,21 @@ export function EmulatorPanel({
       )}
     </PanelSection>
   );
+}
+
+/**
+ * Formats a project ID with a demo prefix if we're in offline mode, or uses the
+ * regular ID if we're in hosting only mode.
+ */
+function getProjectIdForMode(
+  projectId: string | undefined,
+  mode: "all" | "hosting"
+): string {
+  if (!projectId) {
+    return "demo-something";
+  }
+  if (mode === "hosting") {
+    return projectId;
+  }
+  return "demo-" + projectId;
 }