+++ src/commands/ext-install.ts
@@ -27,7 +27,7 @@ import {
 import { getRandomString } from "../extensions/utils";
 import { requirePermissions } from "../requirePermissions";
 import * as utils from "../utils";
-import { track } from "../track";
+import { trackGA4 } from "../track";
 import { confirm } from "../prompt";
 import { Options } from "../options";
 import * as manifest from "../extensions/manifest";
@@ -96,11 +96,18 @@ export const command = new Command("ext:install [extensionName]")
       // Should parse spec locally so we don't need project ID.
       source = await createSourceFromLocation(needProjectId({ projectId }), extensionName);
       await displayExtInfo(extensionName, "", source.spec);
-      void track("Extension Install", "Install by Source", options.interactive ? 1 : 0);
+      void trackGA4("extension_added_to_manifest", {
+        published: "local",
+        interactive: options.nonInteractive ? "false" : "true",
+      });
     } else {
-      void track("Extension Install", "Install by Extension Ref", options.interactive ? 1 : 0);
       extensionName = await canonicalizeRefInput(extensionName);
       extensionVersion = await extensionsApi.getExtensionVersion(extensionName);
+
+      void trackGA4("extension_added_to_manifest", {
+        published: extensionVersion.listing?.state === "APPROVED" ? "published" : "uploaded",
+        interactive: options.nonInteractive ? "false" : "true",
+      });
       await infoExtensionVersion({
         extensionName,
         extensionVersion,