+++ src/deploy/functions/prepare.ts
@@ -457,7 +457,13 @@ export async function loadCodebases(
     await runtimeDelegate.build();
 
     const firebaseEnvs = functionsEnv.loadFirebaseEnvs(firebaseConfig, projectId);
-    wantBuilds[codebase] = await runtimeDelegate.discoverBuild(runtimeConfig, firebaseEnvs);
+    wantBuilds[codebase] = await runtimeDelegate.discoverBuild(runtimeConfig, {
+      ...firebaseEnvs,
+      // Quota project is required when using GCP's Client-based APIs
+      // Some GCP client SDKs, like Vertex AI, requires appropriate quota project setup
+      // in order for .init() calls to succeed.
+      GOOGLE_CLOUD_QUOTA_PROJECT: projectId,
+    });
   }
   return wantBuilds;
 }